# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Morex API",
        "description": "API pour l'application mobile Morex - Gestion des finances personnelles",
        "contact": {
            "name": "Morex Support",
            "email": "support@morex.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/goals": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Liste des objectifs",
                "operationId": "ade9bdb09d2491dd10ebc655489f5688",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des objectifs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Goal"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Goals"
                ],
                "summary": "Créer un objectif",
                "operationId": "4b3554baf009c6109ae16d72cdda0434",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "target_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fonds d'urgence"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "savings",
                                            "debt",
                                            "investment",
                                            "custom"
                                        ]
                                    },
                                    "target_amount": {
                                        "description": "Montant cible en FCFA",
                                        "type": "integer",
                                        "example": 2610000
                                    },
                                    "current_amount": {
                                        "description": "Montant déjà épargné",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "target_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "account_id": {
                                        "description": "Compte associé pour suivi auto",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#8B5CF6"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "target"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Objectif créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Goal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/goals/{id}": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Détail d'un objectif",
                "operationId": "285cd4baa464b1f4579b7ecfc29d9c84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'objectif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Goal"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Objectif non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Goals"
                ],
                "summary": "Modifier un objectif",
                "operationId": "b2ea42b020bd44f268427651d3f7c190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "savings",
                                            "debt",
                                            "investment",
                                            "custom"
                                        ]
                                    },
                                    "target_amount": {
                                        "type": "integer"
                                    },
                                    "current_amount": {
                                        "type": "integer"
                                    },
                                    "target_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "account_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "completed",
                                            "cancelled"
                                        ]
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Objectif modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Goal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Objectif non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Goals"
                ],
                "summary": "Supprimer un objectif",
                "operationId": "6345c58cfdd94d7b8fae3e8b9c7d33cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Objectif supprimé"
                    },
                    "404": {
                        "description": "Objectif non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/goals/{id}/contribute": {
            "post": {
                "tags": [
                    "Goals"
                ],
                "summary": "Ajouter une contribution à un objectif",
                "operationId": "ec6d6c02a3e408a5609437439051923f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant à ajouter en FCFA",
                                        "type": "integer",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contribution ajoutée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Goal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Objectif non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Account": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Compte principal"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "bank",
                            "mobile_money",
                            "savings"
                        ]
                    },
                    "initial_balance": {
                        "type": "integer",
                        "example": 100000
                    },
                    "balance": {
                        "type": "integer",
                        "example": 150000
                    },
                    "formatted_balance": {
                        "type": "string",
                        "example": "150 000 FCFA"
                    },
                    "color": {
                        "type": "string",
                        "example": "#3B82F6"
                    },
                    "icon": {
                        "type": "string",
                        "example": "wallet"
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "order_index": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Goal": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Fonds d'urgence"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "savings",
                            "debt",
                            "investment",
                            "custom"
                        ]
                    },
                    "target_amount": {
                        "type": "integer",
                        "example": 2610000
                    },
                    "current_amount": {
                        "type": "integer",
                        "example": 500000
                    },
                    "progress_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 19.2
                    },
                    "formatted_target_amount": {
                        "type": "string",
                        "example": "2 610 000 FCFA"
                    },
                    "formatted_current_amount": {
                        "type": "string",
                        "example": "500 000 FCFA"
                    },
                    "target_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "color": {
                        "type": "string",
                        "example": "#8B5CF6"
                    },
                    "icon": {
                        "type": "string",
                        "example": "target"
                    },
                    "account": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Account"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Sanctum. Utilisez le token reçu lors du login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}